BuildAGame
Whole game works now on Floats instead of INT - so SmartEnemy now is better in following Player

+++ src/com/beginning/main/BasicEnemy.java
@@ -11,13 +11,13 @@ public class BasicEnemy extends GameObject {
 
     private Handler handler;
 
-    public BasicEnemy(int x, int y, ID id, Handler handler) {
+    public BasicEnemy(float x, float y, ID id, Handler handler) {
         super(x, y, id);
 
         this.handler = handler;
 
-        velX = 5;
-        velY = 5;
+        velX = 5.0f;
+        velY = 5.0f;
 
     }
 
@@ -37,18 +37,18 @@ public class BasicEnemy extends GameObject {
             velY *= -1;
         }
 
-        handler.addObject(new Trail(x, y, Color.red, 16, 16, 0.025f, ID.Trail, handler));
+        handler.addObject(new Trail(x, y, Color.red, 16.0f, 16.0f, 0.025f, ID.Trail, handler));
     }
 
     @Override
     public void render(Graphics g) {
         g.setColor(Color.red);  //set color of render object
-        g.fillRect(x, y, 16, 16);  //fill renctangle - Dimenstion of him is 16x16px
+        g.fillRect((int)x, (int)y, 16, 16);  //fill renctangle - Dimenstion of him is 16x16px
     }
 
     @Override
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 16, 16);  //16 x 16 is dimensions of our rectangle
+        return new Rectangle((int)x, (int)y, 16, 16);  //16 x 16 is dimensions of our rectangle
     }
 
 }
